Players should only be able to use a charter if they're in the correct world.
In 1.8 players in creative are able to spawn in any item with any nbt
tags they want, including charters that are filled out with any players
they want.  This commit makes it so players can't create a clan using
whateverrandom users they want without their permission in a
non-creative world.